requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
llama-index-core>=0.12.0
llama-index-embeddings-cohere
llama-index-postprocessor-cohere-rerank
llama-index-vector-stores-qdrant
llama-index-retrievers-bm25
llama-index-llms-gemini
rank_bm25
qdrant-client>=1.7.0
langchain-text-splitters
pandas
openpyxl
beautifulsoup4
tqdm
redis>=5.0.0